Added basic layout of critical path test
This test gives a 50/50 chance of creating a user who wants to
join the site and another who is only checking out a few stocks
without logging in. The user who creates an account does so and
then favorites a few stocks.

There might be some issues with the dynamic variables being set
to the correct authentication_token when trying to do any sort of
POST request.